Reshuffeled some functions, in order to free fSpec of recursive definition, without creating cycles in the import structure of modules.
